Created Repository files and modified services and controllers

+++ backend/backend_springboot/src/main/java/it/unito/iumtweb/springboot/actors/ActorsRepository.java
@@ -1,10 +1,9 @@
 package it.unito.iumtweb.springboot.actors;
 
-import it.unito.iumtweb.springboot.actors.Actors;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-
 @Repository
-public interface ActorsRepository extends JpaRepository<Actors, Integer> {
+public interface ActorsRepository extends JpaRepository<Actors, Long> {
+    // aggiungere metodi custom se necessario (es. findByName)
 }
